+++ app/views/inc/example_foot.php
@@ -1,8 +1,1 @@
-</div>
-<footer>
-    <div style="flex: 1; text-align: left;">
-        <p style="margin-left: 20px;">© <?= date('Y'); ?></p>
-    </div>
-</footer>
-</body>
-</html>
+<?php require __DIR__ . '/foot.php'; ?>