+++ app/views/inc/example_nav.php
@@ -1,10 +1,1 @@
-<nav>
-  <div class="nav-container">
-    <a href="/"><img alt="Site Name" src="/assets/icons/icon.svg" height="80" width="80"></a>
-    <ul>
-      <li><a href="/">Home</a></li>
-      <li><a href="/page_one">One</a></li>
-      <li><a href="/page_two">Two</a></li>
-    </ul>
-  </div>
-</nav>
+<?php require __DIR__ . '/nav.php'; ?>